+++ bmc-opsi/src/main/java/com/oracle/bmc/opsi/model/EnterpriseManagerBridgeSummary.java
@@ -71,6 +71,15 @@ public class EnterpriseManagerBridgeSummary {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("objectStorageBucketStatusDetails")
+        private String objectStorageBucketStatusDetails;
+
+        public Builder objectStorageBucketStatusDetails(String objectStorageBucketStatusDetails) {
+            this.objectStorageBucketStatusDetails = objectStorageBucketStatusDetails;
+            this.__explicitlySet__.add("objectStorageBucketStatusDetails");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("freeformTags")
         private java.util.Map<String, String> freeformTags;
 
@@ -146,6 +155,7 @@ public class EnterpriseManagerBridgeSummary {
                             displayName,
                             objectStorageNamespaceName,
                             objectStorageBucketName,
+                            objectStorageBucketStatusDetails,
                             freeformTags,
                             definedTags,
                             systemTags,
@@ -165,6 +175,8 @@ public class EnterpriseManagerBridgeSummary {
                             .displayName(o.getDisplayName())
                             .objectStorageNamespaceName(o.getObjectStorageNamespaceName())
                             .objectStorageBucketName(o.getObjectStorageBucketName())
+                            .objectStorageBucketStatusDetails(
+                                    o.getObjectStorageBucketStatusDetails())
                             .freeformTags(o.getFreeformTags())
                             .definedTags(o.getDefinedTags())
                             .systemTags(o.getSystemTags())
@@ -214,6 +226,12 @@ public class EnterpriseManagerBridgeSummary {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("objectStorageBucketName")
     String objectStorageBucketName;
+
+    /**
+     * A message describing status of the object storage bucket of this resource. For example, it can be used to provide actionable information about the permission and content validity of the bucket.
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("objectStorageBucketStatusDetails")
+    String objectStorageBucketStatusDetails;
 
     /**
      * Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.