+++ bmc-tenantmanagercontrolplane/src/main/java/com/oracle/bmc/tenantmanagercontrolplane/model/OperationType.java
@@ -13,6 +13,10 @@ public enum OperationType {
     CreateSenderInvitation("CREATE_SENDER_INVITATION"),
     AcceptRecipientInvitation("ACCEPT_RECIPIENT_INVITATION"),
     CancelSenderInvitation("CANCEL_SENDER_INVITATION"),
+    ActivateOrderExistingTenancy("ACTIVATE_ORDER_EXISTING_TENANCY"),
+    RegisterDomain("REGISTER_DOMAIN"),
+    ReleaseDomain("RELEASE_DOMAIN"),
+    CreateChildTenancy("CREATE_CHILD_TENANCY"),
 
     /**
      * This value is used if a service returns a value for this enum that is not recognized by this